PYTHON-525 Fix MongoClient.is_mongos.
It had checked if the cluster type is "sharded" which is only true when doing mongos high-availability with multiple seeds. The cluster type is "single" if we are directly connected to any server, including a mongos.

Thus the proper way to implement is_mongos is to see if the current writable server is a mongos or not.